improve fs cleanup on ulta commit_hash:2ac63c848ce6c0a599391f37559aa4eb3c030536

+++ ulta/service/filesystem_cleanup.py
@@ -3,6 +3,7 @@ import os
 import shutil
 import time
 from pathlib import Path
+from typing import Callable
 
 from yandextank.contrib.netort.netort.resource import ResourceManager
 
@@ -33,12 +34,28 @@ class FilesystemCleanup:
         self._stpd_cache_ttl = stpd_cache_ttl
         self._netort_cache_ttl = netort_cache_ttl
 
-        self._forbiden_dirs = {self._stpd_cache_dir, self._fs.tests_dir, self._fs.tmp_dir}
+        self._forbiden_dirs: set[Path] = {
+            self._stpd_cache_dir,
+            self._fs.tests_dir,
+            self._fs.tmp_dir,
+            self._fs.tests_dir / 'lunapark',
+        }
         if self._job.test_data_dir:
             self._forbiden_dirs.add(Path(self._job.test_data_dir))
             self._forbiden_dirs.add(Path(self._job.test_data_dir.replace('/test_data_', '/')))
         if self._job.artifact_dir_path:
             self._forbiden_dirs.add(Path(self._job.artifact_dir_path))
+        self._forbiden_dirs = {d.resolve() for d in self._forbiden_dirs if d.exists()}
+
+    @staticmethod
+    def _log_errors(function: Callable[['FilesystemCleanup'], None]) -> Callable[['FilesystemCleanup'], None]:
+        def inner(self: 'FilesystemCleanup'):
+            try:
+                function(self)
+            except Exception as e:
+                self._logger.error('error during cleanup %(func)s: {ex}', {'func': function.__name__, 'ex': str(e)})
+
+        return inner
 
     def _get_job_disk_limit(self) -> int:
         if rc := self._job.get_plugins(JobPluginType.RESOURCE_CHECK):
@@ -63,9 +80,17 @@ class FilesystemCleanup:
             dict(stage=stage, path=path, free_space=self._get_free_space(path)),
         )
 
+    def _is_forbiden_dir(self, path: Path) -> bool:
+        return path.resolve() in self._forbiden_dirs
+
+    @_log_errors
     def _cleanup_temporary_dir(self):
+        if not self._fs.tmp_dir.exists():
+            self._logger.debug('temporary folder is not found')
+            return
+
         self._log_free_space('before', self._fs.tmp_dir)
-        tmp_objects = tuple(f for f in self._fs.tmp_dir.iterdir() if f not in self._forbiden_dirs)
+        tmp_objects = tuple(f for f in self._fs.tmp_dir.iterdir() if not self._is_forbiden_dir(f))
         for f in tmp_objects:
             if f.is_dir():
                 shutil.rmtree(f)
@@ -73,6 +98,7 @@ class FilesystemCleanup:
                 f.unlink()
         self._log_free_space('after', self._fs.tmp_dir)
 
+    @_log_errors
     def _remove_old_tests_dirs(self):
         if not self._fs.tests_dir.exists():
             self._logger.debug('tests folder is not found')
@@ -83,7 +109,7 @@ class FilesystemCleanup:
             [
                 (f, f.stat().st_ctime)
                 for f in self._fs.tests_dir.iterdir()
-                if f not in self._forbiden_dirs and f.name != 'stpd-cache' and f.is_dir()
+                if not self._is_forbiden_dir(f) and f.name != 'stpd-cache' and f.is_dir()
             ],
             key=lambda f: f[1],
         )
@@ -93,6 +119,7 @@ class FilesystemCleanup:
             shutil.rmtree(f)
         self._log_free_space('after', self._fs.tests_dir)
 
+    @_log_errors
     def _remove_old_stpd_cache_files(self):
         if not self._stpd_cache_dir.exists():
             self._logger.debug('stpd cache folder is not found')
@@ -100,7 +127,11 @@ class FilesystemCleanup:
 
         self._log_free_space('before', self._stpd_cache_dir)
         stpd_cache = sorted(
-            [(f, f.stat().st_ctime) for f in self._stpd_cache_dir.iterdir() if f not in self._forbiden_dirs],
+            [
+                (f, f.stat().st_ctime)
+                for f in self._stpd_cache_dir.iterdir()
+                if not self._is_forbiden_dir(f) and f.is_file()
+            ],
             key=lambda f: f[1],
         )
 
@@ -111,12 +142,17 @@ class FilesystemCleanup:
             f.unlink()
         self._log_free_space('after', self._stpd_cache_dir)
 
+    @_log_errors
     def _clean_netort_resources(self):
         if self._resource_manager is None:
             self._logger.debug('resource manager is not set')
             return
 
         netort_dir = Path(self._resource_manager.tmp_path_prefix)
+        if not netort_dir.exists():
+            self._logger.debug('netort forlder is not exists')
+            return
+
         self._log_free_space('before', netort_dir)
         netort_dirs = set(
             Path(op.factory.keywords['path_provider'].prefix)
@@ -124,7 +160,7 @@ class FilesystemCleanup:
             if hasattr(op.factory, 'keywords') and 'path_provider' in op.factory.keywords
         )
         netort_objects = sorted(
-            [(f, f.stat().st_ctime) for d in netort_dirs for f in d.rglob('*') if f.is_file()],
+            [(f, f.stat().st_ctime) for d in netort_dirs if d.exists() for f in d.rglob('*') if f.is_file()],
             key=lambda f: f[1],
         )
